move large slow test to test:slow
# Conflicts:
#	test.js
